Add `Hamburger` and `MSCANBlock`
