+++ batchflow/models/torch/blocks/__init__.py
@@ -1,6 +1,7 @@
 """ Blocks: large parts that implement idea/named entity from popular articles. """
 from .core import Block, DefaultBlock, Downsample, Upsample
-from .named_blocks import VGGBlock, ResBlock, ResNeStBlock, DenseBlock, MBConvBlock, InvResBlock, ConvNeXtBlock
+from .named_blocks import (VGGBlock, ResBlock, ResNeStBlock, DenseBlock, MBConvBlock, InvResBlock,
+                           ConvNeXtBlock, MSCANBlock)
 from .attention import SEBlock, SCSEBlock, SimpleSelfAttention, BAM, CBAM, FPA, SelectiveKernelConv, SplitAttentionConv
 from .pyramid import PyramidPooling, ASPP, KSAC
 from ..layers import Combine, Branch # convenience
