+++ batchflow/models/torch/layers/__init__.py
@@ -5,9 +5,10 @@ from .core import Flatten, Dense, DenseAlongAxis, Dropout, AlphaDropout
 from .activation import Activation, RadixSoftmax
 from .normalization import Normalization, LayerNorm
 from .conv import (Conv, ConvTranspose,
-                   DepthwiseConv, DepthwiseConvTranspose, SeparableConv, SeparableConvTranspose,
-                   MultiKernelConv, SharedKernelConv, AvgPoolConv, BilinearConvTranspose)
+                   DepthwiseConv, DepthwiseConvTranspose, SeparableConv, SeparableConvTranspose)
+from .conv_complex import MultiKernelConv, SharedKernelConv, AvgPoolConv, BilinearConvTranspose, MultiScaleConv
 from .pooling import MaxPool, AvgPool, GlobalMaxPool, GlobalAvgPool, ChannelPool
 from .resize import IncreaseDim, Reshape, Crop, Interpolate
 from .combine import Combine
+from .hamburger import Hamburger
 from .wrapper_letters import Branch, AttentionWrapper
