+++ batchflow/models/torch/layers/multilayer.py
@@ -7,13 +7,14 @@ from torch import nn
 from .core import Dense, DenseAlongAxis, Dropout, AlphaDropout
 from .normalization import Normalization
 from .conv import (Conv, ConvTranspose,
-                   DepthwiseConv, DepthwiseConvTranspose, SeparableConv, SeparableConvTranspose,
-                   MultiKernelConv, SharedKernelConv, AvgPoolConv, BilinearConvTranspose)
+                   DepthwiseConv, DepthwiseConvTranspose, SeparableConv, SeparableConvTranspose)
+from .conv_complex import MultiKernelConv, SharedKernelConv, AvgPoolConv, BilinearConvTranspose, MultiScaleConv
 from .pooling import AvgPool, MaxPool, GlobalAvgPool, GlobalMaxPool
 from .resize import IncreaseDim, Reshape, Interpolate
 from .activation import Activation
 from .combine import Combine
 from .pixel_shuffle import PixelShuffle, PixelUnshuffle
+from .hamburger import Hamburger
 from .wrapper_letters import Branch, AttentionWrapper
 
 from ..repr_mixin import ModuleDictReprMixin
@@ -85,6 +86,7 @@ class MultiLayer(ModuleDictReprMixin, nn.ModuleDict):
         'K': 'multi_kernel_conv',
         'q': 'avg_pool_conv',
         'Q': 'bilinear_conv_transpose',
+        'm': 'multi_scale_conv',
 
         # Downsample / upsample
         'v': 'avg_pool',
@@ -110,6 +112,7 @@ class MultiLayer(ModuleDictReprMixin, nn.ModuleDict):
 
         # Wrapper
         'S': 'self_attention',
+        'H': 'hamburger',
     }
 
     LAYERS_MODULES = {
@@ -130,6 +133,7 @@ class MultiLayer(ModuleDictReprMixin, nn.ModuleDict):
         'shared_kernel_conv': SharedKernelConv,
         'avg_pool_conv': AvgPoolConv,
         'bilinear_conv_transpose': BilinearConvTranspose,
+        'multi_scale_conv': MultiScaleConv,
 
         'avg_pool': AvgPool,
         'max_pool': MaxPool,
@@ -146,13 +150,14 @@ class MultiLayer(ModuleDictReprMixin, nn.ModuleDict):
         'branch': Branch,
         'branch_end': Combine,
         'self_attention': AttentionWrapper,
+        'hamburger': Hamburger,
     }
 
     DEFAULT_LETTERS = LETTERS_LAYERS.keys()
     LETTERS_GROUPS = dict(zip(DEFAULT_LETTERS, DEFAULT_LETTERS))
     LETTERS_GROUPS.update({
         'C': 'c', 't': 'c', 'T': 'c', 'w': 'c', 'W': 'c',
-        'k': 'c', 'K': 'c', 'q': 'c', 'Q': 'c',
+        'k': 'c', 'K': 'c', 'q': 'c', 'Q': 'c', 'm': 'c',
         'v': 'p',
         'V': 'P',
         'D': 'd',
@@ -202,7 +207,7 @@ class MultiLayer(ModuleDictReprMixin, nn.ModuleDict):
         non_ascii_letters = set(self.layout) - set(only_ascii_letters)
         if non_ascii_letters:
             raise ValueError(f'Layout `{self.layout}` contains non ASCII letters {non_ascii_letters}!')
-        self.layout = self.layout.replace(' ', '')
+        self.layout = self.layout.replace(' ', '').replace('(', '').replace(')', '')
 
         layout_dict = {}
         for letter in self.layout:
